src/main/java/sandbox/xmen/Costume.java
package sandbox.xmen;

public class Costume {
    private String logo; //encapsulation making state of objects
    private String primaryColor;   //using functions to access
    private String secondaryColor;
    private boolean mask;


    public Costume(String logo, String primaryColor,
                    String secondaryColor, boolean mask) {
        this.logo = logo;
        this.primaryColor = primaryColor;
        this.secondaryColor = secondaryColor;
        this.mask= mask;
    }

    public String getLogo() {
        return logo;
    }

    public void setLogo(String logo) {
        this.logo = logo;
    }

    public String getPrimaryColor() {
        return primaryColor;
    }

    public void setPrimaryColor(String primaryColor) {
        this.primaryColor = primaryColor;
    }

    public String getSecondaryColor() {
        return secondaryColor;
    }

    public void setSecondaryColor(String secondaryColor) {
        this.secondaryColor = secondaryColor;
    }

    public boolean isMask() {
        return mask;
    }

    public void setMask(boolean mask) {
        this.mask = mask;
    }
}
